Refactor functions that provide environmental information out of Analytics into its own utility class (IEnvironmentInfoProvider).
