+++ Reqnroll/Analytics/AnalyticsEventProvider.cs
@@ -13,13 +13,13 @@ namespace Reqnroll.Analytics
     public class AnalyticsEventProvider : IAnalyticsEventProvider
     {
         private readonly IUserUniqueIdStore _userUniqueIdStore;
-        private readonly IEnvironmentWrapper _environmentWrapper;
+        private readonly IEnvironmentInfoProvider _environmentInfoProvider;
         private readonly string _unitTestProvider;
 
-        public AnalyticsEventProvider(IUserUniqueIdStore userUniqueIdStore, UnitTestProviderConfiguration unitTestProviderConfiguration, IEnvironmentWrapper environmentWrapper)
+        public AnalyticsEventProvider(IUserUniqueIdStore userUniqueIdStore, UnitTestProviderConfiguration unitTestProviderConfiguration, IEnvironmentInfoProvider environmentInfoProvider)
         {
             _userUniqueIdStore = userUniqueIdStore;
-            _environmentWrapper = environmentWrapper;
+            _environmentInfoProvider = environmentInfoProvider;
             _unitTestProvider = unitTestProviderConfiguration.UnitTestProvider;
         }
 
@@ -27,11 +27,11 @@ namespace Reqnroll.Analytics
         {
             string userId = _userUniqueIdStore.GetUserId();
             string unitTestProvider = _unitTestProvider;
-            string reqnrollVersion = GetReqnrollVersion();
-            string buildServerName = GetBuildServerName();
-            bool isDockerContainer = IsRunningInDockerContainer();
+            string reqnrollVersion = _environmentInfoProvider.GetReqnrollVersion();
+            string buildServerName = _environmentInfoProvider.GetBuildServerName();
+            bool isDockerContainer = _environmentInfoProvider.IsRunningInDockerContainer();
             string hashedAssemblyName = ToSha256(assemblyName);
-            string platform = GetOSPlatform();
+            string platform = _environmentInfoProvider.GetOSPlatform();
             string platformDescription = RuntimeInformation.OSDescription;
 
             var compiledEvent = new ReqnrollProjectCompilingEvent(
@@ -56,13 +56,13 @@ namespace Reqnroll.Analytics
         {
             string userId = _userUniqueIdStore.GetUserId();
             string unitTestProvider = _unitTestProvider;
-            string reqnrollVersion = GetReqnrollVersion();
-            string targetFramework = GetNetCoreVersion() ?? RuntimeInformation.FrameworkDescription;
-            bool isDockerContainer = IsRunningInDockerContainer();
-            string buildServerName = GetBuildServerName();
+            string reqnrollVersion = _environmentInfoProvider.GetReqnrollVersion();
+            string targetFramework = _environmentInfoProvider.GetNetCoreVersion() ?? RuntimeInformation.FrameworkDescription;
+            bool isDockerContainer = _environmentInfoProvider.IsRunningInDockerContainer();
+            string buildServerName = _environmentInfoProvider.GetBuildServerName();
 
             string hashedAssemblyName = ToSha256(testAssemblyName);
-            string platform = GetOSPlatform();
+            string platform = _environmentInfoProvider.GetOSPlatform();
             string platformDescription = RuntimeInformation.OSDescription;
 
             var runningEvent = new ReqnrollProjectRunningEvent(
@@ -80,70 +80,6 @@ namespace Reqnroll.Analytics
             return runningEvent;
         }
 
-        private string GetOSPlatform()
-        {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                return "Windows";
-            }
-
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-            {
-                return "Linux";
-            }
-
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-            {
-                return "OSX";
-            }
-
-            throw new InvalidOperationException("Platform cannot be identified");
-        }
-
-        private readonly Dictionary<string, string> buildServerTypes
-            = new Dictionary<string, string> {
-                { "TF_BUILD","Azure Pipelines"},
-                { "TEAMCITY_VERSION","TeamCity"},
-                { "JENKINS_HOME","Jenkins"},
-                { "GITHUB_ACTIONS","GitHub Actions"},
-                { "GITLAB_CI","GitLab CI/CD"},
-                { "CODEBUILD_BUILD_ID","AWS CodeBuild"},
-                { "TRAVIS","Travis CI"},
-                { "APPVEYOR","AppVeyor"},
-                { "BITBUCKET_BUILD_NUMBER", "Bitbucket Pipelines" },
-                { "bamboo_agentId", "Atlassian Bamboo" },
-                { "CIRCLECI", "CircleCI" },
-                { "GO_PIPELINE_NAME", "GoCD" },
-                { "BUDDY", "Buddy" },
-                { "NEVERCODE", "Nevercode" },
-                { "SEMAPHORE", "SEMAPHORE" },
-                { "BROWSERSTACK_USERNAME", "BrowserStack" },
-                { "CF_BUILD_ID", "Codefresh" },
-                { "TentacleVersion", "Octopus Deploy" },
-
-                { "CI_NAME", "CodeShip" }
-            };
-
-        private string GetBuildServerName()
-        {
-            foreach (var buildServerType in buildServerTypes)
-            {
-                var envVariable = _environmentWrapper.GetEnvironmentVariable(buildServerType.Key);
-                if (envVariable is ISuccess<string>)
-                    return buildServerType.Value;
-            }
-            return null;
-        }
-
-        private bool IsRunningInDockerContainer()
-        {
-            return _environmentWrapper.GetEnvironmentVariable("DOTNET_RUNNING_IN_CONTAINER") is ISuccess<string>;
-        }
-
-        private string GetReqnrollVersion()
-        {
-            return VersionInfo.AssemblyInformationalVersion;
-        }
 
         private string ToSha256(string inputString)
         {
@@ -163,17 +99,5 @@ namespace Reqnroll.Analytics
             return stringBuilder.ToString();
         }
 
-        private string GetNetCoreVersion()
-        {
-            var assembly = typeof(System.Runtime.GCSettings).GetTypeInfo().Assembly;
-            var assemblyPath = assembly.Location.Split(new[] { '/', '\\' }, StringSplitOptions.RemoveEmptyEntries);
-            int netCoreAppIndex = Array.IndexOf(assemblyPath, "Microsoft.NETCore.App");
-            if (netCoreAppIndex > 0 && netCoreAppIndex < assemblyPath.Length - 2)
-            {
-                return assemblyPath[netCoreAppIndex + 1];
-            }
-
-            return null;
-        }
     }
 }
