+++ Reqnroll/EnvironmentAccess/IEnvironmentInfoProvider.cs
@@ -1,0 +1,11 @@
+﻿namespace Reqnroll.EnvironmentAccess
+{
+    public interface IEnvironmentInfoProvider
+    {
+        string GetOSPlatform();
+        string GetBuildServerName();
+        bool IsRunningInDockerContainer();
+        string GetReqnrollVersion();
+        string GetNetCoreVersion();
+    }
+}
