+++ Tests/Reqnroll.RuntimeTests/Analytics/AnalyticsEventProviderTests.cs
@@ -14,12 +14,13 @@ namespace Reqnroll.RuntimeTests.Analytics
         public void Should_return_the_build_server_name_in_Compiling_Event()
         {
             var userUniqueIdStoreMock = new Mock<IUserUniqueIdStore>();
-            var environmentMock = new Mock<IEnvironmentWrapper>();
-            var sut = new AnalyticsEventProvider(userUniqueIdStoreMock.Object, new UnitTestProvider.UnitTestProviderConfiguration(), environmentMock.Object);
-
-            environmentMock
+            var environmentWrapperMock = new Mock<IEnvironmentWrapper>();
+            environmentWrapperMock
                 .Setup(m => m.GetEnvironmentVariable("TF_BUILD"))
                 .Returns(new Success<string>("true"));
+
+            var environmentInfoProvider = new EnvironmentInfoProvider(environmentWrapperMock.Object);
+            var sut = new AnalyticsEventProvider(userUniqueIdStoreMock.Object, new UnitTestProvider.UnitTestProviderConfiguration(), environmentInfoProvider);
 
             var compilingEvent = sut.CreateProjectCompilingEvent(null, null, null, null, null);
             
@@ -30,12 +31,13 @@ namespace Reqnroll.RuntimeTests.Analytics
         public void Should_return_the_build_server_name_in_Running_Event()
         {
             var userUniqueIdStoreMock = new Mock<IUserUniqueIdStore>();
-            var environmentMock = new Mock<IEnvironmentWrapper>();
-            var sut = new AnalyticsEventProvider(userUniqueIdStoreMock.Object, new UnitTestProvider.UnitTestProviderConfiguration(), environmentMock.Object);
-
-            environmentMock
+            var environmentWrapperMock = new Mock<IEnvironmentWrapper>();
+            environmentWrapperMock
                 .Setup(m => m.GetEnvironmentVariable("TEAMCITY_VERSION"))
                 .Returns(new Success<string>("true"));
+            var environmentInfoProvider = new EnvironmentInfoProvider(environmentWrapperMock.Object);
+
+            var sut = new AnalyticsEventProvider(userUniqueIdStoreMock.Object, new UnitTestProvider.UnitTestProviderConfiguration(), environmentInfoProvider);
 
             var compilingEvent = sut.CreateProjectRunningEvent(null);
             
@@ -46,8 +48,10 @@ namespace Reqnroll.RuntimeTests.Analytics
         public void Should_return_null_for_the_build_server_name_when_not_detected()
         {
             var userUniqueIdStoreMock = new Mock<IUserUniqueIdStore>();
-            var environmentMock = new Mock<IEnvironmentWrapper>();
-            var sut = new AnalyticsEventProvider(userUniqueIdStoreMock.Object, new UnitTestProvider.UnitTestProviderConfiguration(), environmentMock.Object);
+            var environmentWrapperMock = new Mock<IEnvironmentWrapper>();
+            var evironmentInfoProvider = new EnvironmentInfoProvider(environmentWrapperMock.Object);
+
+            var sut = new AnalyticsEventProvider(userUniqueIdStoreMock.Object, new UnitTestProvider.UnitTestProviderConfiguration(), evironmentInfoProvider);
 
             var compilingEvent = sut.CreateProjectRunningEvent(null);
